Rename htmlrewrite to httprewrite due to the expanded package capabilities
