+++ internal/cosmetic/injector.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"github.com/ZenPrivacy/zen-desktop/internal/hostmatch"
-	"github.com/ZenPrivacy/zen-desktop/internal/htmlrewrite"
+	"github.com/ZenPrivacy/zen-desktop/internal/httprewrite"
 	"github.com/ZenPrivacy/zen-desktop/internal/logger"
 )
 
@@ -78,7 +78,7 @@ func (inj *Injector) Inject(req *http.Request, res *http.Response) error {
 	// Why append and not prepend?
 	// When multiple CSS rules define an !important property, conflicts are resolved first by specificity and then by the order of the CSS declarations.
 	// Appending ensures our rules take precedence.
-	if err := htmlrewrite.AppendHeadContents(res, ruleInjection.Bytes()); err != nil {
+	if err := httprewrite.AppendHeadContents(res, ruleInjection.Bytes()); err != nil {
 		return fmt.Errorf("append head contents: %w", err)
 	}
