+++ internal/jsrule/injector.go
@@ -8,7 +8,7 @@ import (
 	"regexp"
 
 	"github.com/ZenPrivacy/zen-desktop/internal/hostmatch"
-	"github.com/ZenPrivacy/zen-desktop/internal/htmlrewrite"
+	"github.com/ZenPrivacy/zen-desktop/internal/httprewrite"
 	"github.com/ZenPrivacy/zen-desktop/internal/logger"
 )
 
@@ -73,7 +73,7 @@ func (inj *Injector) Inject(req *http.Request, res *http.Response) error {
 	}
 	injection = append(injection, injectionEnd...)
 
-	if err := htmlrewrite.PrependBodyContents(res, injection); err != nil {
+	if err := httprewrite.PrependBodyContents(res, injection); err != nil {
 		return fmt.Errorf("prepend body contents: %w", err)
 	}
